@@ -4,17 +4,17 @@
   "private": true,
   "dependencies": {
     "@babel/runtime": "7.0.0-beta.55",
-    "axios": "^0.19.0",
-    "bootstrap": "^4.1.3",
-    "firebase-tools": "^6.7.2",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
+    "axios": "^0.19.2",
+    "bootstrap": "^4.4.1",
+    "firebase-tools": "^6.12.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-glitch-effect": "^1.0.4",
     "react-reveal": "^1.2.2",
     "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.5",
-    "react-typist": "^2.0.4",
-    "reactstrap": "^6.4.0"
+    "react-typist": "^2.0.5",
+    "reactstrap": "^6.5.0"
   },
   "scripts": {
     "start": "react-scripts start",
